added use of config file for logs
