+++ dask_io/optimizer/main.py
@@ -4,14 +4,15 @@ import math
 import time
 import datetime 
 import logging
+from logging.config import fileConfig
 
 from dask_io.optimizer.clustered import apply_clustered_strategy
 from dask_io.optimizer.modifiers import get_used_proxies, get_array_block_dims
 from dask_io.utils.utils import LOG_TIME, LOG_DIR
 
-logging.basicConfig(filename=os.path.join(LOG_DIR, LOG_TIME + '.log'), level=logging.DEBUG) # to be set to WARNING
+fileConfig('logging_config.ini')
+logger = logging.getLogger(name).addHandler(logging.NullHandler())
 
-DEBUG_MODE = False
 
 def clustered_optimization(graph):
     """ Applies clustered IO optimization on a Dask graph.
@@ -47,33 +48,5 @@ def optimize_func(dsk, keys):
     t = time.time()
     dask_graph = dsk.dicts
     dask_graph = clustered_optimization(dask_graph)
-    logging.info("Time spent to create the graph: {0:.2f} milliseconds.".format((time.time() - t) * 1000))
-
-    def neat_print_graph(graph, log=True):
-        for k, v in graph.items():
-            if log: 
-                logging.debug(f"\nkey: {k}")
-            else:
-                print(f"\nkey: {k}")
-
-            if isinstance(v, dict):
-                for k2, v2 in v.items():
-                    if log: 
-                        logging.debug(f"\tk: {k2}")
-                        logging.debug(f"\t{v2} \n")
-                    else:
-                        print(f"\tk: {k2}")
-                        print(f"\t{v2} \n")
-
-            else:
-                if log: 
-                    logging.debug(f"\tv: {v}")
-                else:
-                    print(f"\tv: {v}")
-
-    neat_print_graph(dsk, log=True)
-
-    if DEBUG_MODE:
-        raise ValueError("stop here")
-
+    logger.info("Time spent to create the graph: {0:.2f} milliseconds.".format((time.time() - t) * 1000))
     return dsk
